requirements_local.txt
apache-airflow==3.1.5
apache-airflow-providers-amazon==9.18.0
apache-airflow-providers-postgres==6.5.0
apache-airflow-providers-sftp==5.5.0